vhosts

+++ bin/vhost_delete.sh
@@ -1,16 +1,24 @@
 #!/bin/bash
 
-# Create a variable with the sitename.
-echo "Provide the name of the vhost you want to delete, followed by [ENTER]:";
-read vhost;
+if [[ "$1" ]]; then
 
-# Delete all files associated with this site.
-sudo rm -r /var/www/html/vhosts/${vhost};
-sudo rm "/etc/apache2/sites-available/${vhost}.local.conf";
-sudo rm "/etc/apache2/sites-enabled/${vhost}.local.conf";
+	vhost="$1";
 
-# Restart Apache.
-sudo service apache2 restart;
-echo "The ${vhost} vhost was deleted successfully.";
-echo "Apache was restarted. All set.";
-exit 0;
+	if [[ -d "/var/www/html/vhosts/${vhost}" ]]; then
+		# Delete all files associated with this site.
+		sudo rm -r /var/www/html/vhosts/${vhost};
+		sudo rm "/etc/apache2/sites-available/${vhost}.local.conf";
+		sudo rm "/etc/apache2/sites-enabled/${vhost}.local.conf";
+
+		# Restart Apache.
+		sudo service apache2 restart;
+		echo "The ${vhost} vhost was deleted successfully.";
+		echo "Apache was restarted. All set.";
+		exit 0;
+	else
+		echo "No such vhost found.";
+	fi
+else
+	echo "Usage: vhost-delete <hostName>"
+fi
+
